more spritesets, move lemming ex, ratelimit differently, play sound from lemming
